+++ src/7segment/_7segment.cpp
@@ -1,6 +1,6 @@
 #include "7segment/_7segment.hpp"
 #if SEVEN_SEGMENT
-#include "helpers.h"
+#include "helpers/helpers.h"
 
 void _7segment::begin(bool is_cathode)
 {