+++ src/extraHelpers/exthelpers.h
@@ -1,6 +1,0 @@
-#pragma once
-#include "settings.h"
-#include <Arduino.h>
-#include <string.h>
-
-void strncpy_T(char *buf,char str[],int size);